Move cookie_helpers.rb to lib/cdo
