+++ dashboard/config/initializers/session_store.rb
@@ -1,4 +1,4 @@
-require 'cookie_helpers'
+require 'cdo/cookie_helpers'
 session_cookie_key = environment_specific_cookie_name('_learn_session')
 Dashboard::Application.config.session_store :cookie_store,
   key: session_cookie_key,
