+++ lib/cdo/cookie_helpers.rb
